Split crud.py to make linting happy
